LZMR-47 added dependency injection for  SHController constructor
