+++ Netbeans/SuperheroSightings/src/main/java/com/sg/superherosightings/controller/SHController.java
@@ -8,9 +8,10 @@ package com.sg.superherosightings.controller;
 import com.sg.superherosightings.dao.SuperheroSightingsDao;
 import com.sg.superherosightings.dto.Characters;
 import com.sg.superherosightings.dto.Location;
-import com.sg.superherosightings.dto.Organization;
 import com.sg.superherosightings.dto.Sighting;
-import com.sg.superherosightings.service.SHService;
+import com.sg.superherosightings.service.CharactersService;
+import com.sg.superherosightings.service.LocationsService;
+import com.sg.superherosightings.service.SightingsService;
 import java.util.List;
 import javax.inject.Inject;
 import javax.servlet.http.HttpServletRequest;
@@ -26,8 +27,9 @@ import org.springframework.web.bind.annotation.RequestMethod;
 @Controller
 public class SHController {
 
-    SHService service;
-    SuperheroSightingsDao dao;
+    CharactersService charactersService;
+    SightingsService sightingsService;
+    LocationsService locationsService;
     
     /*
     
@@ -36,16 +38,18 @@ public class SHController {
     */
 
     @Inject
-    public SHController(SuperheroSightingsDao dao) {
-        this.dao = dao;
+    public SHController(CharactersService charactersService, SightingsService sightingsService, LocationsService locationsService) {
+        this.charactersService = charactersService;
+        this.sightingsService = sightingsService;
+        this.locationsService = locationsService;
     }
 
     @RequestMapping(value = {"/", "index", ""}, method = RequestMethod.GET)
     public String loadWindow(HttpServletRequest request, Model model) {
-        List<Sighting> sightings = dao.getAllSightings();
+        List<Sighting> sightings = sightingsService.getAllSightings();
         System.out.println(sightings);
-        List<Characters> characters = dao.getAssociatedCharacters(sightings);
-        List<Location> locations = dao.getAssociatedLocations(sightings);
+        List<Characters> characters = charactersService.getAssociatedCharacters(sightings);
+        List<Location> locations = locationsService.getAssociatedLocations(sightings);
         model.addAttribute("sightings", sightings);
         model.addAttribute("locations", locations);
         model.addAttribute("characters", characters);
@@ -66,46 +70,9 @@ public class SHController {
 
     
 
-    @RequestMapping(value = {"/viewLocations"}, method = RequestMethod.GET)
-    public String loadLocations(HttpServletRequest request, Model model) {
-        List<Location> allLocations = dao.getAllLocations();
-        model.addAttribute("locations", allLocations);
-        
-        
-        String display = (request.getParameter("viewType"));
-        model.addAttribute("display", display);
+    
 
-        try {
-            int id = Integer.parseInt(request.getParameter("locationID"));
-            Location location = dao.getLocationById(id);
-            model.addAttribute("location", location);
-        } catch (Exception e) {
-
-        }
-        
-        
-        return "locations";
-    }
-
-    @RequestMapping(value = {"/newLocation"}, method = RequestMethod.POST)
-    public String createLocation(HttpServletRequest request, Model model) {
-        Location myLocation = new Location();
-        myLocation.setLocationName(request.getParameter("locationNameJSP"));
-        myLocation.setDescription(request.getParameter("descriptionNameJSP"));
-        myLocation.setLatitude(Double.parseDouble(request.getParameter("latitudeJSP")));
-        myLocation.setLongitude(Double.parseDouble(request.getParameter("longitudeJSP")));
-        myLocation.setStreetNumber(request.getParameter("streetNumberJSP"));
-        myLocation.setStreetName(request.getParameter("streetNameJSP"));
-        myLocation.setCity(request.getParameter("cityJSP"));
-        myLocation.setState(request.getParameter("stateJSP"));
-        myLocation.setZip(request.getParameter("zipJSP"));
-
-        model.addAttribute("location", request.getParameterValues("locations"));
-        dao.addLocation(myLocation);
-
-        return "redirect:viewLocations";
-
-    }
+
